+++ RedPandaIDE/compiler/compiler.cpp
@@ -37,10 +37,11 @@
 #define COMPILE_PROCESS_END "---//END//----"
 
 Compiler::Compiler(const QString &filename, bool onlyCheckSyntax):
-    QThread(),
-    mOnlyCheckSyntax(onlyCheckSyntax),
-    mFilename(filename),
-    mRebuild(false),
+    QThread{},
+    mOnlyCheckSyntax{onlyCheckSyntax},
+    mFilename{filename},
+    mRebuild{false},
+    mForceEnglishOutput{false},
     mParserForFile()
 {
     getParserForFile(filename);
@@ -685,7 +686,8 @@ void Compiler::runCommand(const QString &cmd, const QString  &arguments, const Q
         }
         env.insert("PATH",path);
     }
-    //env.insert("LANG","en");
+    if (compilerSet() && compilerSet()->forceEnglishOutput())
+        env.insert("LANG","en");
     env.insert("LDFLAGS","-Wl,--stack,12582912");
     env.insert("CFLAGS","");
     env.insert("CXXFLAGS","");