+++ RedPandaIDE/editor.cpp
@@ -1373,7 +1373,6 @@ void Editor::mouseReleaseEvent(QMouseEvent *event)
                     QString filename = mParser->getHeaderFileName(mFilename,s);
                     pMainWindow->openFile(filename);
                     return;
-
                 } else if (mParser->enabled()) {
                     gotoDefinition(p);
                     return;
@@ -4350,14 +4349,13 @@ void Editor::updateFunctionTip(bool showTip)
     // Don't bother scanning the database when there's no identifier to scan for
 
     // Only do the cumbersome list filling when showing a new tooltip...
-
     if (s != pMainWindow->functionTip()->functionFullName()
             && !mParser->parsing()) {
         pMainWindow->functionTip()->clearTips();
         QList<PStatement> statements=mParser->getListOfFunctions(mFilename,
                                                                   s,
                                                                   functionNamePos.line);
-
+//      qDebug()<<"finding function list:"<<s<<" - "<<statements.length();
         foreach (const PStatement statement, statements) {
             pMainWindow->functionTip()->addTip(
                         statement->command,
@@ -4672,7 +4670,9 @@ void Editor::gotoDefinition(const QSynedit::BufferCoord &pos)
         filename = statement->definitionFileName;
         line = statement->definitionLine;
     }
-    Editor *e = pMainWindow->openFile(filename);
+    Editor *e = pMainWindow->editorList()->getOpenedEditorByFilename(filename);
+    if (!e)
+        e = pMainWindow->openFile(filename);
     if (e) {
         e->setCaretPositionAndActivate(line,1);
     }