+++ libs/qsynedit/qsynedit/syntaxer/cpp.h
@@ -23,6 +23,7 @@ namespace QSynedit {
 
 class CppSyntaxer: public Syntaxer
 {
+public:
     enum class TokenId {
         Comment,
         Directive,
@@ -50,10 +51,9 @@ class CppSyntaxer: public Syntaxer
         rsDocstring,
         rsStringEscapeSeq,
         rsRawString, rsSpace,rsRawStringNotEscaping,rsRawStringEnd,rsChar,
-        rsDefineIdentifier, rsDefineRemaining
+        rsDefineIdentifier, rsDefineRemaining,
     };
 
-public:
     explicit CppSyntaxer();
     CppSyntaxer(const CppSyntaxer&)=delete;
     CppSyntaxer operator=(const CppSyntaxer&)=delete;