+++ libs/qsynedit/qsynedit/syntaxer/syntaxer.h
@@ -54,15 +54,11 @@ struct SyntaxState {
     int parenthesisLevel; // current parenthesis embedding level (needed by rainbow color)
 //    int leftBraces; // unpairing left braces in the current line ( needed by block folding)
 //    int rightBraces; // unparing right braces in the current line (needed by block folding)
-    QVector<IndentInfo> indents;
+    QVector<IndentInfo> indents; // indents stack (needed by auto indent)
     IndentInfo lastUnindent;
-//    QVector<int> indents; // indents stack (needed by auto indent)
-//    int firstIndentThisLine; /* index of first indent that appended to the indents
-//                              *  stack at this line ( need by auto indent) */
-//    QVector<int> matchingIndents; /* the indent matched ( and removed )
-//                              but not started at this line
-//                                (need by auto indent) */
     bool hasTrailingSpaces;
+    std::shared_ptr<QVariant> extraData;
+
     bool operator==(const SyntaxState& s2);
     IndentInfo getLastIndent();
     IndentType getLastIndentType();